Add support for ingredients in element mixin
This changes the return value of the getIngredient() method.
If the element has ingredients it will return the ingredient
object instead of a value.

getValue() has been introduced to return the value of an ingredient
and should be used as replacement for getIngredient if a value is
expected.

For compatibility getEssence() returns an ingredient object, but logs
a warning in the console to tell users to use getIngredient instead.